Abstracted VispyrStack into utils
