+++ cdk/utils/generateUserData.ts
@@ -1,8 +1,10 @@
 import { UserData } from 'aws-cdk-lib/aws-ec2';
 
-export const generateUserData = (commands: string[]) => {
+const generateUserData = (commands: string[]) => {
   const userData = UserData.forLinux();
   userData.addCommands(...commands);
 
   return userData;
 };
+
+export default generateUserData;
